pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.5.3",
    "chromadb>=1.3.4",
    "elasticsearch==8.11.1",
    "fastapi>=0.121.2",
    "flower>=2.0.1",
    "gradio>=5.23.1",
    "ipykernel==6.30.1",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "tiktoken>=0.12.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]



[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
